The humble beginnings of DB test suite
The TestDB is a single test entry point that prepares and launches all
actual test calls. It doesn't run at all if the connection to the real
DB was not established (kind of a quick run of 'go test', skips the
expensive bits).

The init() func tries to detect when the tests are being called on a
clean test DB and sets up or skips accordingly.