gorm: migrate comment insertion
Fixup CommentID field to be int64 instead of string, since string can
not be a serial primary key. This might break something with template
rendering, but seems to be OK at the first glance.

Rearrange Comment and Commenter structs a bit to play nicely with gorm.
See comment in data.go for details.